I think this fixes the segfault from #123
I am not going to pretend to fully comprehend what is going on, but this seems to fix the segfault.
There is a chance the segfault is really just a strange surgical linker bug.
There is also a chance of some other form of bugged cffi or memory freeing related to roc allocations.

This at least fixes locally for me and hopefully will fix in ci so we can at least cut a release.
I think this is a case of small changes revealing/hiding a larger bug that is hard to track down.